Added hook to propagate CI flag
Right now, if you create a non-code patch before getting feedback from
jenkins, that patch will never get CI review from it as jenkins will
review only the patchset that has code changes. This hook is triggered
on any comment, and propagates the CI value to the latest patchset if
it does not have a CI value already.

Change-Id: I0b6f3a53edd0a8d2a0114fb728fff0054c11d466
Signed-off-by: David Caro <dcaroest@redhat.com>