Handle ext tracker 'None' status in set_modified hook
The hook checks the external trackers (patches) of the bug,
and according to the external tracker 'status' field it checks if the patch is
open ('POST') or closed ('MERGED').
The problem is when the user adds manual patch by adding the patch number,
the 'status' parameter remain 'None'.

Adding some code to query gerrit (if the 'status' == 'None') to get the
real status of the patch.

Change-Id: I56d37381e7af07a431b69d44b695667ad6a7f9ad
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1235
Signed-off-by: Shlomi Ben-David <sbendavi@redhat.com>