Fix missing patchset
If someone press the 'rebase' button or add a comment it will run the
hook dispatcher and it will take some time (few seconds) to finish,
while during that time the 'rebase' button remains available
and allows to press on it again.
This behaviour cause patches to disappear.

Example:
current patch: #5 ==> clicked on the rebase button
(you should be moved to patch #6) ==> clicked on the rebase button again
(you are now on patch #7) while patch #6 isn't exist.

Added code to catch the 'rebase' or added comments action and skip
the rest of the checks.

In Addition:
- Rewrite the logging logic
- Fixed some pep8 errors
- Removed/Added some function doc strings

Change-Id: Ie797a8dd8e7b5aa1415e668bb430ff65ed62ff52
Jira-ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1192
Signed-off-by: Shlomi Ben-David <sbendavi@redhat.com>